@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Product } from 'src/app/model/product';
 import { ProductBOService } from 'src/app/service/product-bo.service';
+import { NgbModal, NgbModalRef } from '@ng-bootstrap/ng-bootstrap';
 
 @Component({
   selector: 'app-producto-bo',
@@ -10,7 +11,8 @@ import { ProductBOService } from 'src/app/service/product-bo.service';
 export class ProductoBOComponent implements OnInit {
 
   constructor(
-    private productService : ProductBOService
+    private productService : ProductBOService,
+    private modalService: NgbModal
   ) { }
 
   public products: Product[] = [];
@@ -24,6 +26,10 @@ export class ProductoBOComponent implements OnInit {
         console.error(err);
       },
     })
+  }
+
+  openModal(content: any): NgbModalRef {
+    return this.modalService.open(content, { centered: true });
   }
 
   accion(accion: string) {
